Разделяет написанный код на модули
